+++ tests/test_technician_model.py
@@ -3,36 +3,37 @@
 import pytest
 from unittest.mock import Mock
 from ..scheduler.models.technician_model import Technician
-from ..scheduler.models.task_model import Task 
-from ..scheduler.models.base_model import BaseModel
+from ..scheduler.models.task_model import Task
 
-class TestTechnicianModel(BaseModel):
+class TestTechnicianModel:
     @pytest.fixture
-    def db(self):
-        return Mock()
-
+    def technician(self):
+        return Technician(1, 'John Doe', 8, [])
+    
     @pytest.fixture
-    def technician(self, db):
-        return Technician(1, 'John Doe', 8, [], db)
+    def task(self):
+        return Task(1, 'Task 1', 3, 'Test Description 1', status='Pending')
 
     def test_init(self, technician):
         assert technician.id == 1
         assert technician.name == 'John Doe'
-        assert technician.availability == 8
-        assert technician.tasks == []
-        #assert technician.db == db
+        assert technician.assignable.availability == 8
+        assert technician.assignable.tasks == []
 
-    def test_add_task(self, technician):
-        task = Task(1, 'Task 1', 3, 'Test Description 1', self.db)
+    def test_add_task(self, technician, task):
         technician.add_task(task)
-        assert technician.availability == 5
-        assert technician.tasks == [task]
-        #technician.db.update_technician.assert_called_once_with(technician)
+        assert technician.assignable.availability == 5
+        assert technician.assignable.tasks == [task]
 
-    def test_add_task_not_enough_availability(self, technician):
-        task = Task(1, 'Task 1', 10, 'Test Description 1', self.db)  # Task duration is greater than technician's availability
+    def test_add_task_not_enough_availability(self, technician, task):
+        task.duration = 10 # Task duration is greater than technician's assignable
         with pytest.raises(ValueError, match='Not enough availability'):
             technician.add_task(task)
-        assert technician.availability == 8  # Availability should not change
-        assert technician.tasks == []  # No task should be added
-        #technician.db.update_technician.assert_not_called()  # update_technician should not be called
+        assert technician.assignable.availability == 8  # availability should not change
+        assert technician.assignable.tasks == []  # No task should be added
+
+    def test_add_task_already_assigned(self, technician, task):
+        technician.add_task(task)
+        with pytest.raises(ValueError, match='Task already assigned'):
+            technician.add_task(task)
+        assert technician.assignable.availability == 5